% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_carga_datos.R
\name{descargar_leer_estaciones}
\alias{descargar_leer_estaciones}
\title{Carga datos estaciones}
\usage{
descargar_leer_estaciones(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{Un string del id de una estacion meteorologica del INTA}

\item{ruta_archivo}{Un string de la ruta donde se leera el archivo de los datos.}
}
\value{
Un conjunto de datos
}
\description{
Esta funcion sirve para descagar y leer datos de las estaciones meteorologicas que contiene el paquete. Toma el id de una estacion que se quieren leer los datos (como primer argumento) y una ruta de archivo (segundo argumento) para descargar los datos en esa ruta, en caso de que aún no esten descargados allí, y leerlos. Devuelve un conjunto de datos (.csv) de esa estacion específica.
}
\examples{
descargar_leer_estaciones("NH0472", "../estacion2")
}
