% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_grafico_temp_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico de la temperatura abrigo 150cm mensual}
\usage{
grafico_temperatura_mensual(estaciones, colores, titulo = "Temperatura")
}
\arguments{
\item{estaciones}{Un conjunto de datos (csv) con datos de una o más estaciones del INTA}

\item{colores}{String o vector con colores en ingles para cada estacion}

\item{titulo}{String del titulo del gráfico, sino hay uno por defecto.}
}
\value{
Un gráfico de lineas
}
\description{
#' Esta función crea un gráfico de líneas que muestra la temperatura abrigo 150 cm promedio
mensual registrada por cada estación meteorológica. A partir del un conjunto
de datos (primer argumento) calcula el promedio de temperatura por mes y genera
una visualización comparando las estaciones, en caso de haber más de una, con distintos colores.
}
\details{
Si no se proporcionan suficientes colores, la función asigna colores
aleatorios automáticamente
}
\examples{
grafico_temperatura_mensual(etaciones, c('red', 'blue'), "Grafico temperatura mensual por estacion")
}
